end-to-end/requirements.txt
accelerate==1.4.0
datasets==3.3.2
evaluate==0.4.3
hf_transfer==0.1.9
huggingface-hub==0.29.1
hydra-core==1.3.2
kagglehub==0.3.10
kubernetes==32.0.1
lm_eval==0.4.7
matplotlib==3.10.0
metaflow-checkpoint==0.1.8
networkx==3.4.2
nltk==3.9.1
numpy==2.2.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
pandas==2.2.3
pyarrow==19.0.1
pydantic==2.10.6
requests==2.32.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentencepiece==0.2.0
standard-imghdr==3.13.0
tensorboard==2.19.0
tokenizers==0.21.0
torch==2.6.0
torchao==0.8.0
torchdata==0.11.0
transformers==4.49.0
wandb==0.19.7